Refactor Authentication interfaces - Remove Core SDK Authentication inheritance from Widgets Authentication class - Make AuthenticationManager implement Widgets Authentication class - Move public Widgets Authentication class to a supposedly public package
MOB-4206
