+++ app/src/main/java/com/glia/exampleapp/MainFragment.kt
@@ -30,7 +30,7 @@ import com.glia.exampleapp.ExampleAppConfigManager.createDefaultConfig
 import com.glia.exampleapp.Utils.getAuthenticationBehaviorFromPrefs
 import com.glia.widgets.GliaWidgets
 import com.glia.widgets.core.notification.NotificationActionReceiver
-import com.glia.widgets.core.visitor.Authentication
+import com.glia.widgets.authentication.Authentication
 import com.glia.widgets.entrywidget.EntryWidget
 import com.glia.widgets.launcher.EngagementLauncher
 import com.google.android.material.appbar.MaterialToolbar
@@ -248,7 +248,7 @@ class MainFragment : Fragment() {
                 initGliaWidgets()
                 requireActivity().runOnUiThread { showAuthenticationDialog(callback) }
             }
-        } else if (authentication != null && authentication!!.isAuthenticated) {
+        } else if (authentication?.isAuthenticated == true) {
             callback.onAuthenticated()
         } else {
             showAuthenticationDialog(callback)
