+++ widgetssdk/src/main/java/com/glia/widgets/core/authentication/AuthenticationManager.kt
@@ -1,7 +1,7 @@
 package com.glia.widgets.core.authentication
 
 import com.glia.androidsdk.RequestCallback
-import com.glia.widgets.core.visitor.Authentication
+import com.glia.widgets.authentication.Authentication
 import com.glia.widgets.di.Dependencies
 import com.glia.widgets.di.Dependencies.repositoryFactory
 import com.glia.widgets.helper.Logger
@@ -22,7 +22,7 @@ internal class AuthenticationManager(
     override fun authenticate(
         jwtToken: String,
         externalAccessToken: String?,
-        requestCallback: RequestCallback<Void>
+        authCallback: RequestCallback<Void>?
     ) {
         Dependencies.destroyControllersAndResetQueueing()
 
@@ -33,11 +33,11 @@ internal class AuthenticationManager(
                 repositoryFactory.secureConversationsRepository.subscribe()
             }
 
-            requestCallback.onResult(void, gliaException)
+            authCallback?.onResult(void, gliaException)
         }
     }
 
-    override fun deauthenticate(requestCallback: RequestCallback<Void>) {
+    override fun deauthenticate(authCallback: RequestCallback<Void>?) {
         Logger.i(TAG, "Unauthenticate")
         //Need to end engagement before it's done on the core side to prevent unexpected behavior
         Dependencies.destroyControllersAndResetEngagementData()
@@ -46,12 +46,13 @@ internal class AuthenticationManager(
         //and we don't need secure conversations data for un-authenticated visitors.
         repositoryFactory.secureConversationsRepository.unsubscribeAndResetData()
 
-        authentication.deauthenticate(requestCallback)
+        authentication.deauthenticate(authCallback)
     }
 
-    override fun isAuthenticated(): Boolean = authentication.isAuthenticated
+    override val isAuthenticated: Boolean
+        get() = authentication.isAuthenticated
 
-    override fun refresh(jwtToken: String?, externalAccessToken: String?, authCallback: RequestCallback<Void>?) {
+    override fun refresh(jwtToken: String, externalAccessToken: String?, authCallback: RequestCallback<Void>?) {
         Logger.i(TAG, "Refresh authentication")
         authentication.refresh(jwtToken, externalAccessToken, authCallback)
     }
